fix: only log & send base error message
Most (all?) errors that occur during traversal get wrapped in traversal-
related errors so they look like "block not found" or "could not reify" when
they may just be something like the client closing the connection prematurely.

So instead, let's unwrap the errors to find the base one and just send that and
hopefully it's more informative. If it is a "block not found", or other
traversal-related error then that'll be the base.